Mark DefaultActivity in Provision direct boot aware
It can't be a setup wizard without being direct boot aware. It doesn't
store any state so it is also safe to do so.

This bug reveals possibilities where HOME role is different from the
chosen HOME activity, but it is out of the scope of this CL to fix that.

Bug: 355664417
Test: PMS chooses it as the first HOME activity, which matches the
selection of HOME role.
Test: Clank shows "Add to Home screen" with Provision as the OOBE
package.
Flag: EXEMPT bug fix

Change-Id: I9d2505965fdf343468a3eefe228d7e93b13f914b